+++ Source/Bogus.Tests/DataSetTests/InternetTests.cs
@@ -107,6 +107,18 @@ namespace Bogus.Tests.DataSetTests
       }
 
       [Fact]
+      public void can_get_color_in_format()
+      {
+         internet.Color(format:ColorFormat.Rgb).Should().Be("rgb(77,14,104)");
+      }
+
+      [Fact]
+      public void can_get_color_in_grayscale()
+      {
+         internet.Color(grayscale: true).Should().Be("#4d4d4d");
+      }
+
+      [Fact]
       public void can_get_url_with_path()
       {
          internet.UrlWithPath().Should().Be("https://ambrose.net/soft/deposit");