+++ Source/Bogus.Tests/RandomizerTest.cs
@@ -317,6 +317,18 @@ namespace Bogus.Tests
       }
 
       [Fact]
+      public void generate_hash()
+      {
+         r.Hash().Should().Be("91da090b74f2b910be0dd5991af6398351ac2ef3");
+      }
+
+      [Fact]
+      public void generate_small_hash()
+      {
+         r.Hash(20).Should().Be("91da090b74f2b910be0d");
+      }
+
+      [Fact]
       public void random_word_tests()
       {
          //r.Words(3).Should().Be("");