+++ HISTORY.md
@@ -1,3 +1,15 @@
+## v22.0.7
+* New `Address.AreaCircle(lat,lon,rad)` to generate random latitude/longitude GPS points within a specified radius.
+* New `Address.Geohash()` generates a random Geohash.
+* New `Address.Depth()` generates a random depth (down to Mariana Trench).
+* New `Address.Altitude()` generate a random height (up to Mount Everest). 
+* New `Internet.Color()` format options: CSS `rgb(...)` and delimited RGB.
+* New `System.AndroidId()` to generate GCM registration ID.
+* New `System.ApplePushToken()` to generate a random Apple Push Token.
+* New `System.BlackBerryPin()` to generate a random Black Berry PIN.
+* New `Randomizer.Hash()` to generate random hashes of specified length.
+* New `Randomizer.String2()` to generate random strings with specified character sets.
+
 ## v22.0.6
 * Added `Randomizer.String` method to generate strings. Uses `Chars()` method.
 * PR 136: Improve speed of `DataSet.ParseTokens()`. Thanks @danij!